Using correct steam php file
I was looking into  what is causing the the crash on clicking a pub. It
seems that the pub information is not being stored in permanent
storeage.

Possible solution: check out perm storeage is opened in tabview. It may
not be calling .open() or the getSchedule clone may not have access to
it.